pyproject.toml
[project]
name = "mission-control"
version = "0.1.0"
description = "AI-powered Orchestration Platform"
requires-python = ">=3.13"
dependencies = [
    "ansible>=11.5.0",
    "ansible-runner>=2.4.1",
    "bcrypt==3.2.2",
    "beautifulsoup4>=4.14.2",
    "celery>=5.5.1",
    "docker>=7.1.0",
    "fastapi>=0.115.12",
    "fastmcp>=2.12.4",
    "flower>=2.0.1",
    "gitpython>=3.1.45",
    "jinja2>=3.1.6",
    "paramiko>=4.0.0",
    "passlib>=1.7.4",
    "podman>=5.6.0",
    "pydantic>=2.11.3",
    "pymongo>=4.12.0",
    "pyppeteer>=2.0.0",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.0.0",
    "requests>=2.32.5",
    "rx",
    "scapy>=2.6.1",
    "typer>=0.17.4",
    "uuid7>=0.1.0",
    "uvicorn>=0.34.2",
    "xscan",
]

[tool.uv.sources]
xscan = { path = "libs/xscan-1.3.1-py3-none-any.whl" }
rx = { path = "libs/rx-0.1.0-py3-none-any.whl" }

[dependency-groups]
dev = [
    "celery-types>=0.23.0",
]
